added topology class

+++ Components/CurrentSource.cpp
@@ -1,6 +1,6 @@
 #include "CurrentSource.h"
 
-CurrentSource::CurrentSource(string id, map<string, string>& CurrentSourceNetlist, map<string, int>& CurrentSourceValues):TwoTerminal(id , CurrentSourceNetlist , CurrentSourceValues)
+CurrentSource::CurrentSource(string id, map<string, string>& CurrentSourceNetlist, map<string, double>& CurrentSourceValues):TwoTerminal(id , CurrentSourceNetlist , CurrentSourceValues)
 {
 
 }
@@ -12,5 +12,5 @@ string CurrentSource::getType()
 
 void CurrentSource::printComponentInfo()
 {
-	printf("CurrentSource with default value : %d, t1 : %s, t2 : %s\n", this->defaultValue, this->getT1().c_str(), this->getT2().c_str());
+	printf("CurrentSource with default value : %lf, t1 : %s, t2 : %s\n", this->defaultValue, this->getT1().c_str(), this->getT2().c_str());
 }
