+++ Components/NMOS.h
@@ -3,7 +3,7 @@ class NMOS : public ThreeTerminal
 {
 
 public:
-	NMOS(string id , map<string,int>&info , map<string,string>&netlist);
+	NMOS(string id , map<string,double>&info , map<string,string>&netlist);
 	string getType();
 	void printComponentInfo();
 };
