+++ Components/Resistor.cpp
@@ -1,6 +1,6 @@
 #include "Resistor.h"
 
-Resistor::Resistor(string id, map<string, string>& ResistorNetlist, map<string, int>& ResistorValues):TwoTerminal(id , ResistorNetlist , ResistorValues)
+Resistor::Resistor(string id, map<string, string>& ResistorNetlist, map<string, double>& ResistorValues):TwoTerminal(id , ResistorNetlist , ResistorValues)
 {
 
 }
@@ -12,5 +12,5 @@ string Resistor::getType()
 
 void Resistor::printComponentInfo()
 {
-	printf("Resistor with default value : %d, t1 : %s, t2 : %s\n", this->defaultValue, this->getT1().c_str(), this->getT2().c_str());
+	printf("Resistor with default value : %lf, t1 : %s, t2 : %s\n", this->defaultValue, this->getT1().c_str(), this->getT2().c_str());
 }
