+++ Components/Resistor.h
@@ -3,7 +3,7 @@ class Resistor : public TwoTerminal
 {
 private:
 public:
-	Resistor(string id, map<string, string>& ResistorNetlist, map<string, int>& ResistorValues);
+	Resistor(string id, map<string, string>& ResistorNetlist, map<string, double>& ResistorValues);
 	string getType();
 	void printComponentInfo();
 };
