+++ Components/VoltageSource.h
@@ -3,7 +3,7 @@ class VlotageSource : public TwoTerminal
 {
 private:
 public:
-	VlotageSource(string id, map<string, string>& VlotageSourceNetlist, map<string, int>& VlotageSourceValues);
+	VlotageSource(string id, map<string, string>& VlotageSourceNetlist, map<string, double>& VlotageSourceValues);
 	string getType();
 	void printComponentInfo();
 };
